Module solar::reinhart_sky

Structs

ReinhartSky

A structure that helps creating discretized Skies, using Reinhart’s discretization

Constants

TNAZ 🔒

The number of patches in teach of Tregenza’s (i.e., Reinhart’s with MF=1) sky subdivisions. The angle must come in Radians.

Functions

bins_in_row 🔒

Returns the number of elements in row row, starting from 0.

cone_solid_angle 🔒

Calculates the solid angle of a Cone.

patch_solid_angle 🔒

row_height 🔒

Returns the height of a full row (the cap is only half this height)