Struct solar::reinhart_sky::ReinhartSky

pub struct ReinhartSky {
    pub mf: usize,
    n_rows: usize,
    row_max_sin: Vec<f64>,
    acc_bins: Vec<usize>,
    pub n_bins: usize,
}

Fields
mf: usize

Subdivition scheme

n_rows: usize

The number of rows, including Cap

row_max_sin: Vec<f64>

An array with the Sin(max angle) for every row.

This is useful for transforming a vector into a bin, as it avoids the use of slow trigonometric functions.

acc_bins: Vec<usize>

An array with the number of bins accumulated before each row.

n_bins: usize

The number of elements in this sky discretization (including ground)

Implementations

impl ReinhartSky

pub fn n_bins(mf: usize) -> usize

Calculates the number of total bins in a Reinhart’s discretization including the ground

fn bins_row(&self, nbin: usize) -> usize

Calculates the row in which a certain bin is located

pub fn bin_solid_angle(&self, nbin: usize) -> f64

Calculates te solid angle of a certain bin

fn raccum(&self, row: usize) -> usize

The number of sky elements accumulated up to row row (inclusive). Rows are Zero-indexed (i.e., the first one is 0)

fn row_altitude(&self, row: usize) -> (f64, f64)

Returns the minimum and maximum altitude of a row, in Radians

fn sin_altitude_to_row(&self, sin_altitude: f64) -> usize

Gets the sin of an altitude

E.g., Receives the z component of a direction (if the vector is normalized) and returns the row towards which that vector is pointing.

fn xy_to_bin(&self, row: usize, x: f64, y: f64) -> usize

Gets the position of the bin within a row.

E.g., gets the x and y components of a vector pointing to the Sky and returns the bin pointed within that row

pub fn dir_to_bin(&self, dir: Vector3D) -> usize

Returns the bin number pointed by direction dir

pub fn bin_dir(&self, nbin: usize) -> Vector3D

Gets a Vector3D pointing to the centre the bin.

pub fn new(mf: usize) -> Self

Creates a new Reinhart sky discretization