Struct solar::PerezSky

pub struct PerezSky {}

This struct packs functions from two sources:

Perez, R., Ineichen, P., Seals, R., Michalsky, J. and Stewart, R. (1990), “Modeling daylight availability and irradiance components from direct and global irradiance”

Perez, R., R. Seals, and J. Michalsky (1993) All-Weather Model for Sky Luminance Distribution - Preliminary Configuration and Validation I do not have access to the latter, so I borrowed code from Radiance’s gendaymtx.c file

Implementations

impl PerezSky

fn sky_clearness(
diffuse_horizontal_irrad: f64,
direct_normal_irrad: f64,
solar_zenith: f64
) -> f64

Equation 1

fn sky_brightness(
diffuse_horizontal_irrad: f64,
air_mass: f64,
extraterrestrial_irradiance: f64
) -> f64

Equation 2

fn precipitable_water_content(dew_point_temp: f64) -> f64

Equation 3… dew_point_temp in C

fn clearness_category(clearness_index: f64) -> usize

Get the sky clearness category (Table 1 in the 1990 paper)

fn diffuse_illuminance_ratio(
p_water_content: f64,
cos_zenit: f64,
sky_brightness: f64,
index: usize
) -> f64

Calculates the diffuse illuminance/diffuse radiance ratio according to Equation 7 and Table 4 of Perez et al. 1990

fn calc_params(zenith: f64, epsilon: f64, delta: f64) -> [f64; 5]

Calculates the Perez sky params

pub fn get_sky_func_standard_time(
units: SkyUnits,
solar: &Solar,
date: Date,
dew_point: f64,
diffuse_horizontal_irrad: f64,
direct_normal_irrad: f64
) -> Box<dyn Fn(Vector3D) -> f64>

Returns a function that can be used to calculate the Perez’s sky values. The input to the returned function is a Vector3D pointing towards the sky (IT MUST BE NORMALIZED); the returned value is specified by the units parameter.
The Date must be in local time

pub fn gen_sky_vec(
mf: usize,
solar: &Solar,
date: Date,
weather_data: CurrentWeather,
units: SkyUnits
) -> Result<Matrix, String>

pub fn update_sky_vec(
vec: &mut Matrix,
mf: usize,
solar: &Solar,
date: Date,
weather_data: CurrentWeather,
units: SkyUnits
) -> Result<(), String>