Struct solar::Solar

pub struct Solar {
    pub(crate) latitude: f64,
    pub(crate) longitude: f64,
    pub(crate) standard_meridian: f64,
}

The solar equivalent of Date’s “day of the year”. The distinction is there so that we don’t mistake solar and standard time

Fields
latitude: f64

Latitude in Radians. South is negative, North is positive.

longitude: f64

Longitude (in Radians). East is negative, West is positive

Note that this is Radiance’s conventions, which is the opposite of EPW files.

standard_meridian: f64

Standard meridian (in Radians). East is negative, West is positive.

This value is essentially -15.0*TimeZone (e.g., GMT+1 becomes -15.0)

Note that this is Radiance’s conventions, which is the opposite of EPW files.

Implementations

impl Solar

pub fn new(latitude: f64, longitude: f64, standard_meridian: f64) -> Self

Builds a Solar object from a Latitude, Longitude and Standard meridian (in Radians)

pub fn solar_standard_time_difference(&self, n: f64) -> f64

Returns the difference between the solar and the standard time in minutes

pub fn unwrap_solar_time(&self, n: Time) -> f64

Returns the content of a Time enum. Transforms to Solar if the type of the Enum is Standard

pub fn unwrap_standard_time(&self, n: Time) -> f64

Returns the content of a Time enum. Transforms to Standard if the type of the Enum is Solar

pub fn equation_of_time(&self, n: f64) -> f64

The Equation of Time based on the day of year (can have decimals)

n should be in solar time, but this variable does not change daily so it probably does not matter… let’s just treat it as Float

pub fn declination(&self, n: f64) -> f64

Declination (in Radians), according to Equation 1.6.1B

n should be in solar time, but this variable does not change daily so it probably does not matter… let’s just treat it as Float

pub(crate) fn b(&self, n: f64) -> f64

Equation 1.4.2 in the Book.

n should be in solar time, but this variable does not change daily so it probably does not matter… let’s just treat it as Float

pub fn normal_extraterrestrial_radiation(&self, n: f64) -> f64

Normal extraterrestrial radiation (Gon) Equation 1.4.1b from Duffie and Beckman

n should be in solar time, but this variable does not change daily so it probably does not matter… let’s just treat it as Float

pub fn hour_angle(&self, n: Time) -> f64

Returns the hour angle in degrees

pub fn sunrise_sunset(&self, n: f64) -> (Time, Time)

Gets the sunset time (equation 1.6.10) n should be in solar time, but since it does not change much on a daily basis, we treat it as an Float

pub fn sun_position(&self, n: Time) -> Option<Vector3D>

Builds a vector that points towards the sun.

Z is up, Y is North and X is East